extract the logic of filter pipeline processing into a separate module
